sync and clean up
