+++ typify-impl/src/enums.rs
@@ -1454,71 +1454,6 @@ mod tests {
     }
 
     #[test]
-    fn test_enum_default_generation_simple_variant() {
-        let schema_json = r##"
-        {
-            "definitions": {
-                "test-enum": {
-                    "type": "string",
-                    "enum": ["failure", "skipped", "success"],
-                    "default": "failure"
-                }
-            }
-        }
-        "##;
-
-        let schema: RootSchema = serde_json::from_str(schema_json).unwrap();
-
-        let mut type_space = TypeSpace::default();
-        type_space.add_ref_types(schema.definitions).unwrap();
-
-        let type_id = type_space.ref_to_id.get("test-enum").unwrap();
-        let type_entry = type_space.id_to_entry.get(type_id).unwrap();
-
-        let mut output = OutputSpace::default();
-        type_entry.output(&type_space, &mut output);
-        let actual = output.into_stream();
-
-        let expected = quote! {
-        #[derive(Clone, Copy, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
-        pub enum TestEnum {
-            #[serde(rename = "failure")]
-            Failure,
-            #[serde(rename = "skipped")]
-            Skipped,
-            #[serde(rename = "success")]
-            Success,
-        }
-        impl ToString for TestEnum {
-            fn to_string(&self) -> String {
-                match *self {
-                    Self::Failure => "failure".to_string(),
-                    Self::Skipped => "skipped".to_string(),
-                    Self::Success => "success".to_string(),
-                }
-            }
-        }
-        impl std::str::FromStr for TestEnum {
-            type Err = &'static str;
-            fn from_str(value: &str) -> Result<Self, Self::Err> {
-                match value {
-                    "failure" => Ok(Self::Failure),
-                    "skipped" => Ok(Self::Skipped),
-                    "success" => Ok(Self::Success),
-                    _ => Err("invalid value"),
-                }
-            }
-        }
-        impl Default for TestEnum {
-            fn default() -> Self {
-                TestEnum::Failure
-            }
-        }};
-
-        assert_eq!(expected.to_string(), actual.to_string());
-    }
-
-    #[test]
     fn test_tricky_internally_tagged_enum() {
         let schema_json = r##"
         {
